chore: refactors example app structure
