+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.chargebeereactnative">
+          package="com.chargebee.android">
 
 </manifest>
